[sil] Move ValueOwnershipKind computation code from SILValue.cpp -> ValueOwnerhsipKindClassifier.{cpp,h}.
This piece of code is large enough to be its own local compilation unit.
